+++ packages/plugin-packages/mashroom-portal-remote-app-registry-k8s/src/js/jobs/ScanK8SPortalRemoteAppsBackgroundJob.ts
@@ -37,7 +37,7 @@ export default class ScanK8SPortalRemoteAppsBackgroundJob implements ScanBackgro
 
     async _scanKubernetesServices(): Promise<void> {
         context.lastScan = Date.now();
-        context.error = null;
+        context.errors = [];
 
         const namespaces = await this._determineNamespaces();
         context.namespaces = namespaces;
@@ -104,8 +104,8 @@ export default class ScanK8SPortalRemoteAppsBackgroundJob implements ScanBackgro
                     }
                 }
             } catch (error: any) {
-                this._logger.error(`Error during scan of k8s namespace: ${namespace}`, error);
-                context.error = error.message;
+                context.errors.push(`Error scanning services in namespace ${namespace}: ${error.message}`);
+                this._logger.error(`Error scanning services in namespace ${namespace}`, error);
             }
         }
 
@@ -122,8 +122,9 @@ export default class ScanK8SPortalRemoteAppsBackgroundJob implements ScanBackgro
                         namespaces.push(ns.metadata.name);
                     }
                 });
-            } catch (e) {
-                this._logger.error(`Unable to determine namespaces from labelSelector: ${this._k8sNamespacesLabelSelector}`, e);
+            } catch (e: any) {
+                context.errors.push(`Error scanning namespaces in cluster: ${e.message}`);
+                this._logger.error(`Error scanning namespaces with labelSelector: ${this._k8sNamespacesLabelSelector}`, e);
             }
         }
         this._logger.debug('Determined namespaces to scan:', namespaces.join(', '));