+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [unreleased]
 
+## 2.0.4 (Mai 9, 2022)
+
  * Remote App Registry and K8S remote app registry: Show Apps with errors on top of the list in the Admin UI
  * K8S remote app registry: The Admin UI shows now all successfully registered Apps even if scanning some namespaces fails due to
    missing permissions